requirements.txt
gunicorn
flask
torch
torch-geometric
ogb
pandas
numpy
scikit-learn
python-dotenv
gdown
requests